requirements.txt
django>=4.2
python-docx>=1.1.0
langdetect>=1.0.9
gunicorn>=21.2.0
